refactor(admin): extract shared admin primitives and formatters
Move formatNumber/formatDuration and the shared UI primitives
(StatusBadge, RoleBadge, TrendDelta, KpiCard, UserAvatar,
SortableHeader, SearchInput, ExportButton, TimeRangeSelector,
DayOption) out of Admin.tsx into
frontend/src/components/admin/shared/{format.ts,primitives.tsx}
as named exports, and import them back in. Pure move, no logic
changes.
